backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
celery==5.3.1
redis==4.5.1
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
Pillow==10.0.0
torch==2.2.0
transformers==4.35.2
python-multipart==0.0.6
python-Levenshtein==0.21.1
reportlab==4.0.4
pydantic==1.10.10
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0
spacy==3.6.1
fuzzywuzzy==0.18.0
rapidfuzz==3.4.0
pdf2image==1.16.3
